backend/pyproject.toml
[project]
name = "ai-fridge-elf-backend"
version = "0.1.0"
description = "AI Fridge Elf - 數位冰箱管理系統後端服務"
readme = "../README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "AI Fridge Elf Team"}
]
keywords = ["fridge", "food-management", "line-bot", "ai-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: Chinese (Traditional)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]

dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "psycopg2-binary>=2.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "line-bot-sdk>=3.7.0",
    "openai>=1.6.0",
    "cloudinary>=1.36.0",
    "apscheduler>=3.10.0",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.6",
    "tenacity>=8.2.0",
    "pillow>=10.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "black>=23.10.0",
    "mypy>=1.6.0",
]

[project.urls]
Repository = "https://github.com/yourusername/ai-fridge-elf"
Documentation = "https://github.com/yourusername/ai-fridge-elf/tree/main/docs"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
python-version = "3.11"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "W", "I"]
ignore = ["E501"]  # line too long (handled by black)

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
